Update UserDialogService.cs
In my opinion the EditText must not contain more than one line of text. It feels strange to get a new line on hitting enter in a one line input dialog. With this change the EditText doesn't support more than one line.
I'm not sure if this request fits to your idea of dialog's behaviour. I just want to let you know my opinion about it.